1. Many changes. 2. Added filtrating and wizard pages.

+++ Coopter/Coopter/motioncontroller.h
@@ -31,7 +31,7 @@ private:
     int mRotY;
     int mRotZ;
     int mRotAmplitude;
-    int mRotCircle;
+    double mRotCircle;
 
 };
